Create an abstract decorator to extend the other decorators from
